backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.6.0
pypdf2==3.0.1
pymupdf==1.25.2
pillow==11.0.0
python-dotenv==1.0.1
